refactor(CreateCarUseCase):  CarFactory -> CreateCarUseCase로 변경
- CarFactory를 CreateCarUseCase 로 변경

생성하는 로직이 필요해서 Factory라고 만들었지만
Factory의 역할은 여러 형태의 Car의 초기화 방식을 캡슐화하고 관리해서
새로운 형태를 추가할때 Factory만 수정할 수 있기 위한 목적이라 수정
